shuffle some stuff
